Re-licensing under GPL-2.0
This repository started by modifying the code from Elijah1111/PiRadio.
He has since licensed his code under GPL-2.0, so I need to update this
repository to be compatible with that license. It was an oversight for
me to license that code without his permission, which I did in an
attempt to be compatible with a different repository I was planning to\
incorporate.